freq/project/tools/collector/tradingview_collector/requirements.txt
websocket-client>=1.2.0
pandas>=1.3.0
tabulate>=0.8.0
